Separate into modules the logic, view and model for the app
